+++ target_chains/solana/pyth_solana_receiver_sdk/src/lib.rs
@@ -1,5 +1,9 @@
 use {
-    anchor_lang::declare_id,
+    anchor_lang::{
+        declare_id,
+        prelude::*,
+    },
+    pythnet_sdk::wire::v1::MerklePriceUpdate,
     solana_program::{
         pubkey,
         pubkey::Pubkey,
@@ -9,7 +13,6 @@ use {
 pub mod config;
 pub mod cpi;
 pub mod error;
-pub mod params;
 pub mod pda;
 pub mod price_update;
 pub mod program;
@@ -17,3 +20,16 @@ pub mod program;
 declare_id!("rec5EKMGg6MxZYaMdyBfgwp4d5rB9T1VQH5pJv5LtFJ");
 
 pub const PYTH_PUSH_ORACLE_ID: Pubkey = pubkey!("pythWSnswVUd12oZpeFP8e9CVaEqJg25g1Vtc2biRsT");
+
+#[derive(Debug, AnchorSerialize, AnchorDeserialize, Clone)]
+pub struct PostUpdateAtomicParams {
+    pub vaa:                 Vec<u8>,
+    pub merkle_price_update: MerklePriceUpdate,
+    pub treasury_id:         u8,
+}
+
+#[derive(Debug, AnchorSerialize, AnchorDeserialize, Clone)]
+pub struct PostUpdateParams {
+    pub merkle_price_update: MerklePriceUpdate,
+    pub treasury_id:         u8,
+}
